fix: properly check for error
test_from_file_failed() checks that a missing netrc file results in the
proper error, i.e. Error::Io(std::io::ErrorKind::NotFound. It tries to
do so by checking the string representation of the error, which is not
portable in few ways: the error string for ENOENT, and the value of
errno.

Instead, check the error manually, asserting it is the type wanted.